fix(positions): hold a position as the account states it
Positions were whole numbers from the wire to the callback. A fractional
holding therefore read as flat: half a share was no position at all, in what
the callback reported, in the profit and loss computed from it, and in the
guard that decides whether a contract's slot may be handed to another — which
would hand away a slot the account was still holding.

The quantity is now carried as stated, through the engine's own table, the
shared per-slot value and the account row. The shared one keeps its lock-free
slot by holding the bits.

The API surface is unchanged: it was already fractional, and the truncation was
behind it.

+++ src/types.rs
@@ -1250,7 +1250,10 @@ pub struct AccountState {
 #[derive(Debug, Clone, Default)]
 pub struct PositionInfo {
     pub con_id: i64,
-    pub position: i64,
+    /// The holding exactly as the account states it. Fractional: a holding of
+    /// half a share is a holding, and rounding it to a whole number reported
+    /// it as flat.
+    pub position: f64,
     pub avg_cost: Price,      // per-share avg cost * PRICE_SCALE
     pub symbol: String,
     pub sec_type: String,
